+++ ChatkitiOSExample/ChatkitiOSExample/ViewController.swift
@@ -59,8 +59,8 @@ class ViewController: UIViewController {
 //                currentUser.sendMessage(
 //                    roomId: currentUser.rooms.last!.id,
 //                    text: "Just a message with an attachment",
-//                    attachmentType: .fileURL(imageURL, name: "cucas.jpg")
-////                    attachmentType: .link("https://i.giphy.com/RpByGPT5VlZiE.gif", type: "image")
+//                    attachment: .fileURL(imageURL, name: "cucas.jpg")
+////                    attachment: .link("https://i.giphy.com/RpByGPT5VlZiE.gif", type: "image")
 //                ) { messageId, err in
 //                    guard err == nil else {
 //                        print("Error sending message \(err!.localizedDescription)")