+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - All `PCMessage`s must have a `text` property
 - `subscribeToRoom` now has a (required) completion handler of type
   `PCErrorCompletionHandler`
+- `sendMessage` paramer `attachmentType` renamed to `attachment`
 
 ## [0.8.4](https://github.com/pusher/chatkit-swift/compare/0.8.3...0.8.4) - 2018-05-26